Pre-playtest Version. Before Pause Menu
-Added Splashscreens
-Fixed UI bugs
-Added simple win lose printfs
-Added 2 working levels
-Changed the exit symbol color
-Fixed LineOfSight Accuracy
-Grayed out Options & Credits buttons
-Added kenny

+++ GAM Test/Codes/Array Reader.h
@@ -5,7 +5,7 @@
 
 void arrayReader_init();
 
-void arrayReader_setMap(short size);
+void arrayReader_setMap(int size);
 
 void arrayReader_draw();
